+++ R/helper.R
@@ -15,3 +15,9 @@ measures_to_codomain = function(measures) {
     ParamDbl$new(id = s$id, tags = ifelse(s$minimize, "minimize", "maximize"))
   }))
 }
+
+extract_benchmark_result_learners = function(bmr) {
+  unlist(map(seq_len(bmr$n_resample_results), function(n) {
+    bmr$resample_result(n)$learners
+  }))
+}